Moved update user info to header to always allow to show it. Fixed issue with disconnect not redirecting
