+++ src/projects/initializer.js
@@ -8,7 +8,7 @@ export function initProject(dispatch) {
   return new Promise((resolve, reject) => {
     const projectUrl = projectsActions.getProjectFromUrl();
     // Project url might be /sign-in in the case of sign in (Before user auth)
-    if (!projectUrl || projectUrl === 'sign-in') {
+    if (!projectUrl || projectUrl === 'sign-in' || projectUrl === 'me') {
       return resolve();
     }
     firebaseDb.collection('projects').doc(projectUrl).get().then(snapshot => {
