+++ src/views/pages/tasks/tasks-page.js
@@ -19,10 +19,7 @@ import { firebaseConfig } from 'src/firebase/config';
 import { getUrlSearchParams } from 'src/utils/browser-utils.js';
 import i18n from '../../../i18n.js';
 import './tasks-page.css';
-import { SetUserInfo } from "../../components/set-user-info";
-import { updateUserData } from "src/auth/auth";
-import {getCookie, setCookie} from "../../../utils/browser-utils";
-import getRandomImage from 'src/utils/unsplash';
+import {setCookie} from "../../../utils/browser-utils";
 
 export class TasksPage extends Component {
   constructor() {
@@ -35,7 +32,6 @@ export class TasksPage extends Component {
     this.goToTask = this.goToTask.bind(this);
     this.onLabelChanged = this.onLabelChanged.bind(this);
     this.onNewTaskAdded = this.onNewTaskAdded.bind(this);
-    this.updateUserInfo = this.updateUserInfo.bind(this);
     this.submitNewTask = this.submitNewTask.bind(this);
     this.createTask = this.createTask.bind(this);
 
@@ -48,7 +44,6 @@ export class TasksPage extends Component {
       labels: null,
       labelPool: {},
       isLoadedComments: false,
-      showSetUserInfoScreen: true,
       isCurrentTaskValid: false,
     };
 
@@ -94,13 +89,11 @@ export class TasksPage extends Component {
         search: firebaseConfig.defaultPageToLoad
       })
     }
-
-    this.showSetUserInfo();
   }
 
   setProjectCookie(projectUrl) {
     // Since we are parsing the url we might get undefined as a string
-    if(projectUrl && projectUrl !== 'undefined') {
+    if(projectUrl && projectUrl !== 'undefined' && projectUrl !== 'me') {
       setCookie('project', projectUrl);
     }
   }
@@ -361,45 +354,8 @@ export class TasksPage extends Component {
     this.props.history.push(taskParameter);
   }
 
-  showSetUserInfo() {
-    this.setState({showSetUserInfoScreen: !this.props.auth.isEmailConfigured})
-  }
-
   onLabelChanged(labels) {
     this.setState({labels});
-  }
-
-  renderSetUserInfo() {
-    return (
-      <div>
-        <SetUserInfo
-          isOpen = { (this.state.showSetUserInfoScreen) || this.props.auth.shouldShowUpdateProfile}
-          userInfo={ this.props.auth }
-          photoURL={ this.props.auth.photoURL || getRandomImage()}
-          updateUserInfo={ this.updateUserInfo }
-          onClosed = { () => {
-            this.setState({showSetUserInfoScreen: false});
-            this.props.isShowUpdateProfile(false);
-            this.setState({showSetUserInfoScreen: false})
-          }
-          } />
-      </div>
-    );
-  }
-
-  updateUserInfo(userInfo) {
-    const projectCookie = getCookie('project');
-
-    const oldUserData = this.props.auth;
-    const newUserData = {};
-    newUserData.uid = oldUserData.id;
-    newUserData.email = userInfo.email;
-    newUserData.isEmailConfigured = true; //This is the flag that specify that this module should not show anymore
-    newUserData.displayName = userInfo.name;
-    newUserData.photoURL = userInfo.photoURL;
-    // Update the default project on first time setting user info
-    newUserData.defaultProject = projectCookie || this.props.selectedProject;
-    updateUserData(newUserData);
   }
 
   renderTaskView() {
@@ -451,7 +407,7 @@ export class TasksPage extends Component {
     const isNewTask = this.props.match && this.props.match.params && this.props.match.params.id === 'new-task';
     const isLoading = (!this.state.tasks || (this.props.tasks.size <= 0 && !isNewTask));
     const projectUrl = this.props.match.params.projectUrl;
-    
+
     return (
       <div>
         <div className="g-col">
@@ -487,7 +443,6 @@ export class TasksPage extends Component {
             <div className='task-view-bottom-loader'>&nbsp;</div>: ''
           }
         </div>
-        { this.renderSetUserInfo() }
       </div>
     );
   }
